Cell list pop + code struct rework.
