+++ board.h
@@ -1,4 +1,5 @@
 #include "cells.h"
+#include <stdbool.h>
 
 typedef struct Board
 {
@@ -8,5 +9,8 @@ typedef struct Board
 } board;
 
 board *initBoard(int rows, int cols);
+bool isInBoard(board *pboard, int row, int col);
+cellType getStatus(board *pboard, int row, int col);
+cellList *getOrthogonalCells(board *pboard, int row, int col);
 void floodBoard(board *pboard);
 void printBoard(board *board);
